Suggested changes: rename methods, MockRouter in it's own file
